Fix SQL injection vulnerabilities in `IN` clauses by using parameterized bindings
