+++ src/database/utils.ts
@@ -1,0 +1,15 @@
+/**
+ * Helper to build an IN clause with parameterized bindings.
+ * Ensures protection against SQL injection.
+ * @param ids Array of ID strings
+ * @param baseBinds Base bind variables (e.g. project, tenantId)
+ */
+export function buildInClause(ids: string[], baseBinds: Record<string, unknown> = {}): { clause: string; binds: Record<string, unknown> } {
+  const binds = { ...baseBinds };
+  if (ids.length === 0) {
+    return { clause: "NULL", binds };
+  }
+  const clause = ids.map((_, i) => `:id${i}`).join(",");
+  ids.forEach((id, i) => { binds[`id${i}`] = String(id); });
+  return { clause, binds };
+}
